Отделение TaskContoller в отдельный компонент
